requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   typer
    #   uvicorn
databricks-sql-connector==4.0.5
    # via databricks-sqlalchemy
databricks-sqlalchemy==2.0.7
    # via databricks (pyproject.toml)
et-xmlfile==2.0.0
    # via openpyxl
exceptiongroup==1.3.0
    # via anyio
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   databricks (pyproject.toml)
    #   mcp
httpx-sse==0.4.1
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jsonschema==4.25.0
    # via mcp
jsonschema-specifications==2025.4.1
    # via jsonschema
lz4==4.4.4
    # via databricks-sql-connector
markdown-it-py==4.0.0
    # via rich
mcp==1.13.0
    # via databricks (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
numpy==2.2.6
    # via pandas
oauthlib==3.3.1
    # via databricks-sql-connector
openpyxl==3.1.5
    # via databricks-sql-connector
pandas==2.2.3
    # via databricks-sql-connector
pyarrow==21.0.0
    # via databricks-sqlalchemy
pydantic==2.11.7
    # via
    #   mcp
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   databricks-sql-connector
    #   pandas
python-dotenv==1.1.1
    # via
    #   mcp
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pytz==2025.2
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via databricks-sql-connector
rich==14.1.0
    # via typer
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   python-dateutil
    #   thrift
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   databricks (pyproject.toml)
    #   databricks-sqlalchemy
    #   sqlmodel
sqlmodel==0.0.24
    # via databricks (pyproject.toml)
sse-starlette==3.0.2
    # via mcp
starlette==0.47.2
    # via mcp
thrift==0.20.0
    # via databricks-sql-connector
typer==0.16.0
    # via mcp
typing-extensions==4.14.1
    # via
    #   anyio
    #   exceptiongroup
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   databricks-sql-connector
    #   requests
uvicorn==0.35.0
    # via mcp
